+++ src/main/java/net/silentchaos512/scalinghealth/item/HeartCrystal.java
@@ -38,7 +38,6 @@ public class HeartCrystal extends StatBoosterItem {
     @Override
     boolean isStatIncreaseAllowed(PlayerEntity player, IPlayerData data) {
         //if this would increase the health by 0, we wont't process it
-        //TODO not sure if I clamp the hp. If the increase is of 5 and max hp is of 41, it could be possible to get 45 hp. (have to check)
         return SHItems.heartCrystalIncreaseAmount(player) > 0 || data.getExtraHearts() < SHPlayers.maxHeartCrystals(player) || SHPlayers.hpDisabledByXp(player.world);
     }